Fix refactoring header cart
