+++ src/components/layouts/header/nav-buttons/data/navButtons.types.ts
@@ -1,0 +1,5 @@
+export type HeaderStyleProps = {
+  isHeaderDense: boolean;
+};
+
+export type ShoppingCartDisplay = 'block' | 'none';
