+++ src/components/layouts/header/nav-buttons/shopping-cart-layer-popup/data/index.ts
@@ -1,0 +1,2 @@
+export { default as useShoppingCartLayerPopup } from './useShoppingCartLayerPopup';
+export * from './shoppingCartLayerPopup.types';
