+++ src/components/layouts/header/nav-buttons/shopping-cart-layer-popup/data/shoppingCartLayerPopup.types.ts
@@ -1,0 +1,5 @@
+import { CartAddModalForm } from '@components/main/product-list/cart-add-modal/data';
+
+export type ShoppingCartLayerPopupProps = {
+  cartProductList: CartAddModalForm[];
+};
