+++ src/components/layouts/header/nav-buttons/shopping-cart-layer-popup/shopping-cart-product/data/shoppingCartProduct.types.ts
@@ -1,0 +1,7 @@
+import { CartAddModalForm } from '@components/main/product-list/cart-add-modal/data';
+
+export type ShoppingCartProductProps = {
+  cartProduct: CartAddModalForm;
+  index: number;
+  handleCheckedToggle: (index: number) => void;
+};
