+++ src/components/main/product-list/cart-add-modal/quantity-by-size-form/data/index.ts
@@ -1,1 +1,2 @@
 export { default as useQuantityBySizeForm } from './useQuantityBySizeForm';
+export * from './quantityBySizeForm.types';
